refactor(query): align strict decimal consumers
