+++ src/services/query_service/app/domain/__init__.py
@@ -1,0 +1,1 @@
+"""Framework-independent Query Service policies and value semantics."""
