+++ src/services/query_service/app/services/cash_movement_service.py
@@ -5,9 +5,9 @@ from portfolio_common.reconciliation_quality import COMPLETE
 from portfolio_common.source_data_product_metadata import source_data_product_runtime_metadata
 from sqlalchemy.ext.asyncio import AsyncSession
 
+from ..domain.strict_decimal import decimal_or_zero
 from ..dtos.cash_movement_dto import CashMovementBucket, PortfolioCashMovementSummaryResponse
 from ..repositories.cashflow_repository import CashflowRepository
-from .decimal_amounts import decimal_or_zero
 
 MAX_CASH_MOVEMENT_WINDOW_DAYS = 366
